+++ internal/utils/log/styles.go
@@ -1,6 +1,44 @@
 package log
 
-import "github.com/charmbracelet/lipgloss"
+import (
+	"log/slog"
+	"strings"
+	"sync"
+	"sync/atomic"
+
+	"github.com/charmbracelet/lipgloss"
+	charmlog "github.com/charmbracelet/log"
+)
+
+var (
+	// singleton instances
+	defaultStylesOnce sync.Once
+	defaultStyles     atomic.Pointer[Styles]
+	defaultLoggerOnce sync.Once
+	defaultLogger     atomic.Pointer[slog.Logger]
+)
+
+// initializeStyles creates and initializes the default styles
+func initializeStyles() *Styles {
+	styles := charmlog.DefaultStyles()
+	for _, ls := range levelStyles {
+		levelStr := strings.ToUpper(LogLevelString(ls.level))
+		if len(levelStr) < ls.maxWidth {
+			levelStr = levelStr + strings.Repeat(" ", ls.maxWidth-len(levelStr))
+		}
+		styles.Levels[ls.level] = ls.style.SetString(levelStr)
+	}
+	return styles
+}
+
+// DefaultStyles returns the initialized styles with all levels including Important
+func DefaultStyles() *Styles {
+	defaultStylesOnce.Do(func() {
+		styles := initializeStyles()
+		defaultStyles.Store(styles)
+	})
+	return defaultStyles.Load()
+}
 
 var (
 	debugStyle = lipgloss.NewStyle().Foreground(lipgloss.Color("#808080")) // Gray
